requirements.txt
asgiref==3.4.1
Django==3.0
django-autoslug==1.9.8
django-mptt==0.13.4
django-shortcuts==1.6
mysqlclient==2.1.0
Pillow==8.4.0
waitress==2.0.0
whitenoise==5.3.0
django-heroku==0.3.1
gunicorn==20.1.0
dj-database-url
psycopg2==2.8.6
django-ckeditor==6.2.0